overwrite OCI source label in image
Currently, this label is set to
`https://github.com/chainguard-images/images/tree/main/images/static`
(because the ko base image is this chainguard image).

This leads to issues with dependency managers like renovate, which uses
this label to parse the change log.

Setting this label to the correct source URL fixes such problems.

More Information on those labels:

https://github.com/opencontainers/image-spec/blob/main/annotations.md#pre-defined-annotation-keys

https://github.com/ko-build/ko/issues/1395

https://docs.renovatebot.com/modules/datasource/docker/#description
Signed-off-by: Florian Ströger <florian@florianstroeger.com>